requirements.txt
kalibr
anthropic
openai
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
apify-client

# dev
pytest
